Rewrite of CDAL
Middle of a complete rewrite of the CDAL class. It now uses its own
special Nodes with their own head and tail indices. Only current working
functions are push front and push back. Push front does not shift all
the elements 'up' the list, but rather just places the element in front
of the list, allocating a new node. In this sense, the head index will
not always be 0 for a Node's array.