SDAL updates
SDAL works completely though still needs to be revised.
-added member functions
-fixed contains function like in the other lists
-pos checking for insert, remove, replace, and item_at
-updated clear method so it actually works now

Still need to revise code and remove any references to head since head
is always 0. Tail needs to be reworked to be 0 when list is empty and
point to 1 past current size. User input array size and new array
alocation when sized is reached still not implemented.