SSLL updates after test release
Updated SSLL as necessary to pass the test released here:
https://github.com/scottliu77/cop3530project1testsample.

-renamed LinkedList.h to List.h. Similarly, all references to Linked
List renamed to just List.
-fixed some virtual and override stuff that I still don't fully
understand. At least List.h is entirely virtual now. Other list classes
will need fixes to accomodate this.
-updated some methods to use size_t instead of int
-fixed that weird bug after consecutive insertions (tail pointer wasn't
being updated correctly)
-commented out print statements.
-print method now feeds into ostream instead of cout
-all methods relying on pos now start at 0
-contains method now uses constant references in the equals function
(this is what the test did, so I did it)
-haven't checked with valgrind but there shouldn't be any mem leaks.

With that, SSLL is seemingly perfectly working and would in theory be
ready to submit for a grade *shudders*.